# CHANGE Array to Set for better performance