--- a/.travis.yml
+++ b/.travis.yml
@@ -11,3 +11,4 @@ matrix:
 script:
   - npm run lint
   - npm test
+  - npm run performance