I switched out the Array to a `Set` which improves performance by 1/4. This is especially better when the skipWords-list gets very large like it did on my project.

before (with array):
4605ms
4607ms

after (wit Set):
3737ms
3698ms

I also added the performance-run to travis so we know that it will always work.