@aotaduy 2018-02-16
Wow, Very nice.

I will be trying to review and merge all of this changes today algong with a new release for npm. 
Regards!